backend/requirements.txt
ace_tools==0.0
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
algoliasearch==4.11.2
algoliasearch-django==4.0.0
amqp==5.2.0
annotated-types==0.6.0
anthropic==0.45.2
anyascii==0.3.2
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.35.20
botocore==1.35.20
cachetools==5.5.1
celery==5.4.0
certifi==2025.8.3
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cohere==5.13.11
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
constantly==23.10.4
crispy-bootstrap4==2024.1
cron-descriptor==1.4.3
cryptography==45.0.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-config-url==0.1.1
dj-database-url==2.1.0
dj-rest-auth==7.0.1
Django==5.0.11
django-tenants==3.10.0
django-allauth==65.8.1
django-cacheops==7.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.9.0
django-cotton==0.9.26
django-crispy-forms==2.1
django-eventstream==5.3.2
django-extensions==3.2.3
django-filter==24.2
django-formtools==2.5.1
django-grip==3.5.2
django-htmx==1.19.0
django-htmx-datatables==0.1.0a1
django-humanize==0.1.2
django-mcp-server==0.5.5
django-modelcluster==6.3
django-on-heroku==1.1.2
django-permissionedforms==0.1
django-qrcode==0.3
django-render-block==0.9.2
django-sslserver==0.22
django-storages==1.14.4
django-taggit==5.0.1
django-template-partials==24.4
django-timezone-field==6.1.0
django-treebeard==4.7.1
django-viewflow==2.0.2
django-widget-tweaks==1.5.0
django_csp==3.8
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
draftjs-exporter==5.0.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
elevenlabs==1.2.2
et-xmlfile==1.1.0
etelemetry==0.3.1
eval_type_backport==0.2.2
executing==2.0.1
fastavro==1.10.0
feedparser==6.0.11
filelock==3.14.0
filetype==1.2.0
flower==2.0.1
frozenlist==1.5.0
fsspec==2024.12.0
funcy==2.0
google-auth==2.38.0
googlemaps==4.10.0
greenlet==3.1.1
griffe==1.5.6
gripcontrol==4.2.0
groq==0.16.0
gunicorn==23.0.0
h11==0.16.0
h2==4.1.0
hf-xet==1.1.5
hpack==4.0.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.1
humanize==4.9.0
Hypercorn==0.17.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.7
IMAPClient==3.0.1
incremental==22.10.0
inflection==0.5.1
ipython==8.24.0
isodate==0.6.1
jedi==0.19.1
jiter==0.12.0
jmespath==1.0.1
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyboard==0.13.5
kombu==5.3.7
l18n==2021.3
laces==0.1.1
llvmlite==0.43.0
logfire-api==3.4.0
logtail-python==0.3.3
loguru==0.7.3
looseversion==1.3.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mcp==1.23.0
mdurl==0.1.2
mercadopago==2.3.0
mistralai==1.5.0
msal==1.34.0
msgpack==1.0.8
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==2.21.0
openai-agents==0.9.0
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
pgvector==0.3.2
phonenumbers==8.13.37
pillow==10.3.0
pillow_heif==0.18.0
pipdeptree==2.30.0
playwright==1.48.0
plotly==5.24.1
priority==2.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.2.1
prov==2.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pubcontrol==3.5.0
pure-eval==0.2.2
puremagic==1.30
pyarrow==22.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.12.4
pydantic-graph==0.0.21
pydantic-settings==2.9.1
pydantic_core==2.41.5
pydot==2.0.0
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pynndescent==0.5.13
pyOpenSSL==25.1.0
pyparsing==3.1.2
pypdfium2==5.3.0
pypng==0.20220715.0
PySocks==1.7.1
pytesseract==0.3.13
python-calamine==0.4.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.20
python3-openid==3.2.0
pytrends==4.9.2
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
qrcode==7.4.2
rdflib==7.0.0
redis==5.0.4
referencing==0.36.2
regex==2024.11.6
reportlab==4.2.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.5.3
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.26.1
service-identity==24.1.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.0
sse-starlette==2.3.4
stack-data==0.6.3
starlette==0.46.2
telepath==0.3.1
tenacity==9.0.0
tesseract==0.1.3
threadpoolctl==3.5.0
tokenizers==0.21.0
tornado==6.5.2
tqdm==4.66.4
traitlets==5.14.3
traits==6.3.2
transformers==4.53.0
trio==0.27.0
trio-websocket==0.11.1
tweepy==4.15.0
Twisted==24.3.0
txaio==23.1.1
typer==0.15.4
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
ua-parser==0.18.0
umap-learn==0.5.6
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.34.0
uvloop==0.19.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
whitenoise==6.6.0
Willow==1.8.0
WooCommerce==3.0.0
wsproto==1.2.0
xlrd==2.0.2
yarl==1.18.3
zope.interface==6.4.post2
